requirements.txt
flask==3.1.0
google-generativeai==0.8.4
llama-index-core==0.12.27
PyPDF2==3.0.1
python-dotenv==1.1.0
pymupdf==1.22.5
transformers==4.50.3
torch==2.6.0